feat: move Claude model logic to claude backend and remove model tiers
Remove the ModelTier system (premium/balanced/economy) and tier-based
sorting/inference from the model registry. The core registry is now a
clean, provider-agnostic store — no tier heuristics, no tier-based
fallback chains, no description keyword sniffing.

Move Claude-specific model helpers (supports1mContext, get1mContextModelId)
from core/models.ts to backend/claude-sdk/models.ts where they belong.
The core registry stores capabilities as data; the claude backend owns
the interpretation.

- Remove ModelTier type, TIER_ORDER, tier field from ModelInfo
- Remove inferTierFromText, buildTierByFamily from claude backend
- Remove tier-based sorting from getModels() and convertSdkModels()
- Remove tier-based fallback chain assignment
- Simplify getDefaultModel() to return first registered model
- Move supports1mContext() and get1mContextModelId() to claude backend
