+++ src/core/models.ts
@@ -9,8 +9,6 @@
 
 // ── Types ────────────────────────────────────────────────────────────────────
 
-export type ModelTier = "premium" | "balanced" | "economy";
-
 export type ModelCapabilities = {
   /** Whether the model supports the 1M token context window. */
   supports1mContext: boolean;
@@ -31,18 +29,8 @@ export type ModelInfo = {
   provider: string;
   /** Model capabilities used for backend configuration. */
   capabilities: ModelCapabilities;
-  /** Tier for UI grouping and fallback ordering. */
-  tier: ModelTier;
   /** Model to fall back to on overload/timeout. */
   fallback?: string;
-};
-
-// ── Tier sort order ─────────────────────────────────────────────────────────
-
-const TIER_ORDER: Record<ModelTier, number> = {
-  premium: 0,
-  balanced: 1,
-  economy: 2,
 };
 
 // ── Registry state ──────────────────────────────────────────────────────────
@@ -107,13 +95,13 @@ export function getModel(id: string): ModelInfo | undefined {
   return models.get(id);
 }
 
-/** List all registered models, optionally filtered by provider. Sorted by tier. */
+/** List all registered models, optionally filtered by provider. Returned in registration order. */
 export function getModels(provider?: string): ModelInfo[] {
-  let result = [...models.values()];
+  const result = [...models.values()];
   if (provider) {
-    result = result.filter((m) => m.provider === provider);
+    return result.filter((m) => m.provider === provider);
   }
-  return result.sort((a, b) => TIER_ORDER[a.tier] - TIER_ORDER[b.tier]);
+  return result;
 }
 
 /**
@@ -148,28 +136,14 @@ export function getFallbackModel(modelId: string): string | null {
   return resolveModel(modelId)?.fallback ?? null;
 }
 
-/** Check whether a model supports the 1M token context window. */
-export function supports1mContext(modelId: string): boolean {
-  const info = resolveModel(modelId);
-  // Default to true for unknown models (don't restrict capabilities we can't check)
-  return info?.capabilities.supports1mContext ?? true;
-}
-
-/** Resolve the exact 1M-context model ID for a given model, if one exists. */
-export function get1mContextModelId(modelId: string): string | null {
-  return resolveModel(modelId)?.capabilities.oneMillionContextModelId ?? null;
-}
-
 /**
- * Get the default model for a given tier. Returns the first registered model
- * matching the tier, or the first model overall, or the hardcoded fallback.
+ * Get the default model. Returns the first registered model,
+ * or the hardcoded "default" if the registry is still empty.
  */
-export function getDefaultModel(tier: ModelTier = "balanced"): string {
-  const byTier = [...models.values()].find((m) => m.tier === tier);
-  if (byTier) return byTier.id;
+export function getDefaultModel(): string {
   const first = models.values().next();
   if (!first.done) return first.value.id;
-  return "default"; // ultimate fallback if the registry is still empty
+  return "default";
 }
 
 // ── Provider-scoped clearing ────────────────────────────────────────────────
